requirements.txt
alembic>=1.0.10
click>=7.1.2
Flask>=1.0.2
Flask-Admin>=1.5.7
Flask-Login>=0.4.1
Flask-Migrate>=2.5.2
Flask-SQLAlchemy>=2.4.0
Flask-WTF>=0.14.2
itsdangerous>=1.1.0
Jinja2>=2.11.3
Mako>=1.1.4
MarkupSafe>=1.1.1
python-dateutil>=2.8.1
python-editor>=1.0.4
six>=1.15.0
SQLAlchemy>=1.2.15
Werkzeug>=0.14.1
WTForms>=2.2.1
